refactor: extract create_chat_model and validate_tool_names to agent/model_factory.py (#105)
